Misc UBSan annotations
* Some functions accept negative values which you might not expect
  (given names like "depth" on the argument). These are now commented
  and casted to reflect this.
* When a hint value of 0 is provided, the binary search was started at
  the maximum value rather than the middle, by overflowing an int.
  0 is the only currently provided value, so instead it starts in the
  middle of the range when no hint is provided.
* Annotated implicit truncations to make them explicit when repacking
  integers into bytewise arrays.

Bug: 129300035

+++ rpmb_protocol.h
@@ -88,18 +88,18 @@ enum rpmb_result {
 
 static inline struct rpmb_u16 rpmb_u16(uint16_t val) {
     struct rpmb_u16 ret = {{
-            val >> 8,
-            val >> 0,
+            (uint8_t)(val >> 8),
+            (uint8_t)(val >> 0),
     }};
     return ret;
 }
 
 static inline struct rpmb_u32 rpmb_u32(uint32_t val) {
     struct rpmb_u32 ret = {{
-            val >> 24,
-            val >> 16,
-            val >> 8,
-            val >> 0,
+            (uint8_t)(val >> 24),
+            (uint8_t)(val >> 16),
+            (uint8_t)(val >> 8),
+            (uint8_t)(val >> 0),
     }};
     return ret;
 }
